Add more fields to DataStore structures
The DataStoreSearchParam fields are present on structure versions which
appeared first on NEX 3.7. However, I can't define an exact NEX version
for these fields as some games with matching DDL versions have
different structure revisions applied.

Also, the `totalCountField` was probably added in revision 3, which is
later than the `dataTypes` field. This is because Pokémon Rumble World
only has up to revision 2 and only has the `useCache` and `dataTypes`
params.