Rename ranking_mk7 to ranking_legacy
This seems to be a really old version of the Ranking protocol, since
it's also present in Mario Tennis Open.

The original method and structure names has been added using debug
symbols from Mario Kart 7. The methods labeled as unknown are present on
Mario Tennis Open, but not on MK7.