Add Storage Manager protocol from utility_mk7
The Utility protocol isn't present on the debug symbols of Mario Kart 7.
Instead, there is a different protocol called `StorageManager`, which
uses the same protocol ID as `Utility`.

Since these protocols share the same ID and Storage Manager is only
present on old games like MK7 and Mario Tennis Open, check on the
Utility protocol for these games and redirect the RMC packet to this
protocol.

The method names come from the debug symbols for Mario Kart 7.