Add version checks to DataStore update structures
On Mario Kart 7, the data ID is a uint32 and the version is a uint16.
Check for DataStore version >= 3 and choose the size accordingly.

Since we can't get the proper type of a field when encoding to JSON,
define the type when parsing the fields. Using `typeof` doesn't work
because on the case of the version field, the JavaScript type is always
`number`.