Fix recursion of NEX lists
Since the value of a list is an array and not an object, 2D lists
wouldn't work because `serializeRMCBody` doesn't support arrays. To fix
that, make the list parsing its own function and check if the value is a
list.

If it is, create an object with the list type name and value and recurse
on the same function.